Corrected warning and removed the strip option in order to make analysis through Valgrind/Dr.Memory easier

+++ AutoOutOfScope/src/Main.cpp
@@ -10,8 +10,8 @@ int main(int ac, char *av[]) {
 
 void PotentiallyLeakyFunction() {
     float *array = new float[100];
-    printf("array@[0x%08X]\n", array);
-    Auto(printf("Next freeing array@[0x%08X]\n", array);delete [] array;);
+    printf("array@[0x%08X]\n", (unsigned int) array);
+    Auto(printf("Next freeing array@[0x%08X]\n", (unsigned int) array); delete [] array;);
     const char *function = __FUNCTION__;
     Auto(printf("Bye bye from [%s]\n", function););
 }
